only attempt IPT2 panel if installed
